Accommodate for new update

+++ src/bot.js
@@ -12,6 +12,7 @@ const prefix = process.env.BOT_PREFIX;
 
 let userApplyList = {};
 let answerQuestionList = {};
+let revApplicationList = {};
 
 const clientActivity = ['For someone to apply', 'For the prefix: elixir.'];
 let clientActivityNumber = 0;
@@ -45,13 +46,13 @@ client.on('message', async message => {
             }
         } else if(message.content.toLowerCase().startsWith(prefix)){ 
             if(message.channel.id === process.env.APPLYSEND_CHANNEL_ID){
-                reviewApplication.reviewapply(message, args, Discord, prefix);
+                reviewApplication.reviewapply(message, args, Discord, prefix, revApplicationList);
             } else if(message.channel.id === process.env.QUESTION_CHANNEL){
                 answerQuestion.aswQuestion(args, message, Discord, answerQuestionList);
             } else if(message.channel.type === "dm"){
                 question.askQuestion(message, args, Discord, client);
             } 
-        } else if((message.channel.id === process.env.APPLYSEND_CHANNEL_ID || message.channel.id === process.env.QUESTION_CHANNEL) && !answerQuestionList[message.author.id]){
+        } else if((message.channel.id === process.env.APPLYSEND_CHANNEL_ID || message.channel.id === process.env.QUESTION_CHANNEL) && !answerQuestionList[message.author.id] && !revApplicationList[message.author.id]){
             message.delete();
             const reply = await message.reply(`Please do not talk here! This is only for accepting or rejecting applications or answering question! Go to <#${process.env.OFFICER_CHANNEL_ID}> for discussions!`)
                 .then(setTimeout(() => reply.delete(), 20000));
